minor bug fix - save all three timeseries plots
Before, all three plots were saved to the same path, resulting in overwriting the file and only saving the last plot. By adding the parameter power to the plot_opt_flex_timeseries function definitions and passing the power level as an argument when calling the function in run_ev_case_study - the power level can now be added to the figure path. Therefore all three figures are now saved, as the power level is included in the figure filename.
The new parameter was included in the two function docstrings.